Omit `V1` from `ApiGatewayProxyHandler`
